+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Delegates participant creation to Experiment `create_participant` method and
   `participant_constructor` attribute to allow experiments to specify custom
   Participant classes.
+- Add extensible actions to the dashboard database view.
 
 ## [v-6.4.0](https://github.com/dallinger/dallinger/tree/6.4.0) (2020-08003)
 - Bugfix: Fixes for Dashboard monitor layout and color issues